split utils.js. Allow API access from html player
